Moved Clock class into lib along with expander lib. Minor fixes in clock divider expander.
